Implement REST API to get the satisfaction rate information.
